+++ src/gotenberg_client/_types_compat.py
@@ -4,7 +4,7 @@
 
 import sys
 
-if sys.version_info >= (3, 11):
+if sys.version_info >= (3, 11):  # pragma: no cover
     from typing import Self
-else:
+else:  # pragma: no cover
     from typing_extensions import Self  # noqa: F401